moved app to the top level
